+++ lib/+mp/sm_lin_constraint.m
@@ -205,7 +205,7 @@ classdef sm_lin_constraint < mp.set_manager_opt_model
                 obj.data.vs = subsasgn(obj.data.vs, sc, vs);
             end
             if ~isempty(obj.cache)  %% clear cache of aggregated params
-                obj.cache = [];
+                obj.clear_cached_params();
             end
         end
 
@@ -579,7 +579,7 @@ classdef sm_lin_constraint < mp.set_manager_opt_model
             end
 
             %% clear cached parameters
-            obj.cache = [];
+            obj.clear_cached_params();
 
             %% update dimensions and indexing, if necessary
             dN = N - N0;