+++ lib/+mp/sm_quad_constraint.m
@@ -264,7 +264,7 @@ classdef sm_quad_constraint < mp.set_manager_opt_model
                 obj.data.vs = subsasgn(obj.data.vs, sc, vs);
             end
             if ~isempty(obj.cache)  %% clear cache of aggregated params
-                obj.cache = [];
+                obj.clear_cached_params();
             end
         end
 
@@ -588,7 +588,7 @@ classdef sm_quad_constraint < mp.set_manager_opt_model
             end
 
             %% clear cached parameters
-            obj.cache = [];
+            obj.clear_cached_params();
 
             %% update dimensions and indexing, if necessary
             dMB = MB - MB0;